Move test to headful.spec
